Move files to correct path
