+++ solutions/dijkstra/elixir/approach4.exs
