+++ solutions/dijkstra/go/approach4.go
