+++ solutions/dijkstra/java/approach4.java
