+++ solutions/dijkstra/ruby/approach4.rb
