+++ solutions/ominous-omino/elixir/approach4.exs
