+++ solutions/ominous-omino/go/approach4.go
